# server/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Annotation API",
    "version": "1.0.0",
    "description": "API documentation for the Annotation project"
  },
  "servers": [
    {
      "url": "http://ic-api.ailabs.io.vn/",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Success" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/admin/user/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get user by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}